+++ components/camel-couchbase/pom.xml
@@ -85,55 +85,14 @@
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
-    <profiles>
-    <!-- activate test if the docker socket file is accessible -->
-    <profile>
-        <id>couchbase-tests-docker-file</id>
-        <activation>
-            <file>
-                <exists>/var/run/docker.sock</exists>
-            </file>
-        </activation>
-        <build>
-            <plugins>
-                <plugin>
-                    <artifactId>maven-surefire-plugin</artifactId>
-                    <configuration>
-                        <skipTests>${skipTests}</skipTests>
-                        <forkCount>1</forkCount>
-                        <systemPropertyVariables>
-                            <visibleassertions.silence>true</visibleassertions.silence>
-                        </systemPropertyVariables>
-                    </configuration>
-                </plugin>
-            </plugins>
-        </build>
-    </profile>
 
-    <!-- activate test if the DOCKER_HOST env var is set -->
-    <profile>
-        <id>couchbase-tests-docker-env</id>
-        <activation>
-            <property>
-                <name>env.DOCKER_HOST</name>
-            </property>
-        </activation>
-        <build>
-            <plugins>
-                <plugin>
-                    <artifactId>maven-surefire-plugin</artifactId>
-                    <configuration>
-                        <skipTests>${skipTests}</skipTests>
-                        <forkCount>1</forkCount>
-                        <systemPropertyVariables>
-                            <visibleassertions.silence>true</visibleassertions.silence>
-                        </systemPropertyVariables>
-                    </configuration>
-                </plugin>
-            </plugins>
-        </build>
-    </profile>
-    </profiles>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 </project>
